requirements.txt
fastapi
uvicorn
streamlit
pillow
numpy
scikit-learn
joblib
requests
tensorflow
imbalanced-learn